+++ Changelog.md
@@ -2,8 +2,9 @@
 
 ## Version (NEXT)
 * improved measurement probabilities calculation for observables
+* improved noise application in noisy simulator
 * fixed bug in eigen-hermitian in fastmath backend
-* sorted eigen decompositions consistently across functions and backends 
+* sorted eigen decompositions consistently across functions and backends
 * added diagonal? to MatrixAlgebra protocol
 * added and improved docstrings
 * updated the tutorial